Cambio en var_globales para tomar long long

+++ segunda_entrega/var_globales.h
@@ -1,6 +1,6 @@
 #define TAM_LEXEMA 200
 
-typedef struct { long codigo;
+typedef struct { long long codigo;
                  char lexema[TAM_LEXEMA];
                }token;
